+++ app/Entity.php
@@ -18,13 +18,17 @@ class Entity extends Model
         'entity_type_id',
         'root_entity_id',
         'name',
-        'lasteditor',
+        'user_id',
         'geodata_id',
     ];
 
     protected $appends = [
         'parentIds',
         'parentNames',
+    ];
+
+    protected $with = [
+        'user',
     ];
 
     protected $searchable = [
@@ -87,7 +91,7 @@ class Entity extends Model
             $entity->{$key} = $value;
         }
         $entity->entity_type_id = $entityTypeId;
-        $entity->lasteditor = $user->name;
+        $entity->user_id = $user->id;
         $entity->save();
 
         // TODO workaround to get all (optional, not part of request) attributes
@@ -116,7 +120,7 @@ class Entity extends Model
                 }
             }
 
-            self::addSerial($entity->id, $s->id, $s->text, $nextValue, $user->name);
+            self::addSerial($entity->id, $s->id, $s->text, $nextValue, $user->id);
         }
 
         $entity->children_count = 0;
@@ -143,7 +147,7 @@ class Entity extends Model
         $hasParent = isset($parent);
         $oldRank = $entity->rank;
         $entity->rank = $rank;
-        $entity->lasteditor = $user->name;
+        $entity->user_id = $user->id;
 
         $query;
         if(isset($entity->root_entity_id)) {
@@ -176,12 +180,12 @@ class Entity extends Model
         $entity->save();
     }
 
-    public static function addSerial($eid, $aid, $text, $ctr, $username) {
+    public static function addSerial($eid, $aid, $text, $ctr, $uid) {
         $av = new AttributeValue();
         $av->entity_id = $eid;
         $av->attribute_id = $aid;
         $av->str_val = sprintf($text, $ctr);
-        $av->lasteditor = $username;
+        $av->user_id = $uid;
         $av->save();
     }
 
@@ -201,12 +205,16 @@ class Entity extends Model
         return $this->belongsTo('App\Entity', 'root_entity_id');
     }
 
+    public function user() {
+        return $this->belongsTo('App\User');
+    }
+
     public function bibliographies() {
         return $this->belongsToMany('App\Bibliography', 'references', 'entity_id', 'bibliography_id')->withPivot('description', 'attribute_id')->orderBy('references.attribute_id')->orderBy('references.bibliography_id');
     }
 
     public function attributes() {
-        return $this->belongsToMany('App\Attribute', 'attribute_values')->withPivot('entity_val', 'str_val', 'int_val', 'dbl_val', 'dt_val', 'certainty', 'certainty_description', 'lasteditor', 'thesaurus_val', 'json_val', 'geography_val')->orderBy('attribute_values.attribute_id');
+        return $this->belongsToMany('App\Attribute', 'attribute_values')->withPivot('entity_val', 'str_val', 'int_val', 'dbl_val', 'dt_val', 'certainty', 'certainty_description', 'user_id', 'thesaurus_val', 'json_val', 'geography_val')->orderBy('attribute_values.attribute_id');
     }
 
     public function files() {