+++ resources/js/App.vue
@@ -103,8 +103,11 @@
                             <i class="fas fa-fw fa-user"></i> {{ $auth.user().name }}
                         </a>
                         <ul class="dropdown-menu dropdown-menu-right" aria-labelledby="user-dropdown">
+                            <router-link :to="{name: 'userprofile'}" class="dropdown-item">
+                                <i class="fas fa-fw fa-id-badge"></i> {{ $t('global.user.profile') }}
+                            </router-link>
                             <router-link :to="{name: 'userpreferences', params: { id: $auth.user().id }}" class="dropdown-item">
-                                <i class="fas fa-fw fa-cog"></i> {{ $t('global.user.settings') }}
+                                <i class="fas fa-fw fa-user-cog"></i> {{ $t('global.user.settings') }}
                             </router-link>
                             <a class="dropdown-item" href="#"
                                 @click="logout">