+++ resources/js/app.js
@@ -70,6 +70,7 @@ import {
     faFileWord,
     faFolder,
     faGlobeAfrica,
+    faIdBadge,
     faInfoCircle,
     faLayerGroup,
     faLightbulb,
@@ -81,6 +82,7 @@ import {
     faMapMarkedAlt,
     faMapMarkerAlt,
     faMicrochip,
+    faMobileAlt,
     faMonument,
     faPalette,
     faPaperPlane,
@@ -127,6 +129,7 @@ import {
     faUnlink,
     faUnlockAlt,
     faUser,
+    faUserCog,
     faUserEdit,
     faUsers,
     faVolumeMute,
@@ -153,6 +156,7 @@ import Users from './components/Users.vue';
 import Roles from './components/Roles.vue';
 import Preferences from './components/Preferences.vue';
 import UserPreferences from './components/UserPreferences.vue';
+import UserProfile from './components/UserProfile.vue';
 const DataModel = () => import(/* webpackChunkName: "group-bib" */ './components/DataModel.vue')
 const DataModelDetailView = () => import(/* webpackChunkName: "group-bib" */ './components/DataModelDetailView.vue')
 
@@ -232,6 +236,7 @@ library.add(
     faFileWord,
     faFolder,
     faGlobeAfrica,
+    faIdBadge,
     faInfoCircle,
     faLayerGroup,
     faLightbulb,
@@ -243,6 +248,7 @@ library.add(
     faMapMarkedAlt,
     faMapMarkerAlt,
     faMicrochip,
+    faMobileAlt,
     faMonument,
     faPalette,
     faPaperPlane,
@@ -290,6 +296,7 @@ library.add(
     faUnlink,
     faUnlockAlt,
     faUser,
+    faUserCog,
     faUserEdit,
     faUsers,
     faVolumeMute,
@@ -535,6 +542,14 @@ const router = new VueRouter({
                 auth: true
             }
         },
+        {
+            path: '/profile',
+            name: 'userprofile',
+            component: UserProfile,
+            meta: {
+                auth: true
+            }
+        },
     ]
 });
 // Workaround to load plugin pages, whose routes
@@ -617,6 +632,7 @@ import AttributeSearch from './components/AttributeSearch.vue';
 import CsvTable from './components/CsvTable.vue';
 
 // Reusable Components
+import UserAvatar from './components/UserAvatar.vue';
 import Attributes from './components/AttributeList.vue';
 import EntityTree from './components/EntityTree.vue';
 import EntityTypes from './components/EntityTypeList.vue';
@@ -636,6 +652,7 @@ Vue.component('entity-type-search', EntityTypeSearch);
 Vue.component('label-search', LabelSearch);
 Vue.component('attribute-search', AttributeSearch);
 Vue.component('csv-table', CsvTable);
+Vue.component('user-avatar', UserAvatar);
 Vue.component('attributes', Attributes);
 Vue.component('entity-tree', EntityTree);
 Vue.component('entity-types', EntityTypes);