+++ tests/Feature/ApiEntityTest.php
@@ -35,7 +35,7 @@ class ApiEntityTest extends TestCase
                 'root_entity_id',
                 'geodata_id',
                 'rank',
-                'lasteditor',
+                'user_id',
                 'created_at',
                 'updated_at',
                 'parentIds',
@@ -63,7 +63,18 @@ class ApiEntityTest extends TestCase
             'root_entity_id' => null,
             'geodata_id' => 2,
             'rank' => 1,
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
+            'user' => [
+                'id' => 1,
+                'name' => "Admin",
+                'nickname' => "admin",
+                'email' => "admin@localhost",
+                'phonenumber' => "+49 1234 12345",
+                'created_at' => "2017-12-20 09:47:36",
+                'updated_at' => "2017-12-20 09:47:36",
+                'avatar' => null,
+                'avatar_url' => null,
+            ],
             'created_at' => '2017-12-20 17:10:34',
             'updated_at' => '2017-12-31 16:10:56',
             'parentIds' => [1],
@@ -123,7 +134,7 @@ class ApiEntityTest extends TestCase
                 'entity_id',
                 'attribute_id',
                 'value',
-                'lasteditor',
+                'user_id',
                 'created_at',
                 'updated_at'
             ]
@@ -158,7 +169,7 @@ class ApiEntityTest extends TestCase
                 'entity_id',
                 'attribute_id',
                 'value',
-                'lasteditor',
+                'user_id',
                 'created_at',
                 'updated_at'
             ]
@@ -208,7 +219,7 @@ class ApiEntityTest extends TestCase
                 'entity_id',
                 'attribute_id',
                 'value',
-                'lasteditor',
+                'user_id',
                 'created_at',
                 'updated_at'
             ]
@@ -278,7 +289,7 @@ class ApiEntityTest extends TestCase
                     'attribute_id',
                     'bibliography_id',
                     'description',
-                    'lasteditor',
+                    'user_id',
                     'created_at',
                     'updated_at',
                     'bibliography'
@@ -293,7 +304,7 @@ class ApiEntityTest extends TestCase
                     'attribute_id' => 15,
                     'bibliography_id' => 1318,
                     'description' => 'See Page 10',
-                    'lasteditor' => 'Admin',
+                    'user_id' => 1,
                     'created_at' => '2019-03-08 13:36:36',
                     'updated_at' => '2019-03-08 13:36:36',
                     'bibliography' => [
@@ -306,7 +317,7 @@ class ApiEntityTest extends TestCase
                     'attribute_id' => 15,
                     'bibliography_id' => 1319,
                     'description' => 'Picture on left side of page 12',
-                    'lasteditor' => 'Admin',
+                    'user_id' => 1,
                     'created_at' => '2019-03-08 13:36:48',
                     'updated_at' => '2019-03-08 13:36:48',
                     'bibliography' => [
@@ -321,7 +332,7 @@ class ApiEntityTest extends TestCase
                     'attribute_id' => 13,
                     'bibliography_id' => 1323,
                     'description' => 'Page 10ff is interesting',
-                    'lasteditor' => 'Admin',
+                    'user_id' => 1,
                     'created_at' => '2019-03-08 13:37:09',
                     'updated_at' => '2019-03-08 13:37:09',
                     'bibliography' => [
@@ -353,7 +364,7 @@ class ApiEntityTest extends TestCase
                 'root_entity_id',
                 'geodata_id',
                 'rank',
-                'lasteditor',
+                'user_id',
                 'created_at',
                 'updated_at',
                 'parentIds'
@@ -401,7 +412,7 @@ class ApiEntityTest extends TestCase
             'root_entity_id',
             'geodata_id',
             'rank',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
             'parentIds'
@@ -443,7 +454,7 @@ class ApiEntityTest extends TestCase
             'root_entity_id',
             'geodata_id',
             'rank',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
             'parentIds'
@@ -484,7 +495,7 @@ class ApiEntityTest extends TestCase
             'attribute_id',
             'bibliography_id',
             'description',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
             'bibliography',
@@ -494,7 +505,7 @@ class ApiEntityTest extends TestCase
             'attribute_id' => 14,
             'bibliography_id' => 1337,
             'description' => 'This is a simple test',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
             'bibliography' => [
                 'id' => 1337,
                 'type' => 'article',
@@ -568,7 +579,7 @@ class ApiEntityTest extends TestCase
             'root_entity_id',
             'geodata_id',
             'rank',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
             'parentIds'
@@ -848,7 +859,7 @@ class ApiEntityTest extends TestCase
             'attribute_id',
             'bibliography_id',
             'description',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
         ]);
@@ -857,7 +868,7 @@ class ApiEntityTest extends TestCase
             'attribute_id' => 15,
             'bibliography_id' => 1319,
             'description' => 'Page 12 was wrong, it is Page 15!',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
         ]);
     }