+++ tests/Unit/BibliographyTest.php
@@ -39,7 +39,7 @@ class BibliographyTest extends TestCase
         ]);
 
         $user = new \stdClass;
-        $user->name = 'Test User';
+        $user->id = 1;
 
         $ranges = array_merge([''], range('a', 'z'), range('A', 'Z'));
         foreach($ranges as $letter) {
@@ -47,7 +47,7 @@ class BibliographyTest extends TestCase
             $b->title = 'Title';
             $b->type = 'article';
             $b->citekey = 'No:1999'.$letter;
-            $b->lasteditor = 'Test User';
+            $b->user_id = 1;
             $b->save();
         }